pyproject.toml
[tool.poetry]
name = "manager-backend"
version = "0.1.0"
description = ""
authors = ["ohnas <ohnas12@gmail.com>"]
readme = "README.md"
packages = [{include = "manager_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
black = "^22.12.0"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.1"
facebook-business = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
